backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sendgrid==6.10.0
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
spacy==3.7.2
textstat==0.7.3
sentence-transformers==2.2.2
scikit-learn==1.3.2
tiktoken==0.7.0
